+++ src/main/java/sg/edu/iss/team6/repository/StudentRepository.java
@@ -1,6 +1,10 @@
 package sg.edu.iss.team6.repository;
 
+import java.util.ArrayList;
+import java.util.function.LongBinaryOperator;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import org.springframework.data.jpa.repository.Query;
 import sg.edu.iss.team6.model.Student;
@@ -9,7 +13,7 @@ import sg.edu.iss.team6.model.User;
 import java.util.ArrayList;
 
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
 
     Student findByStudentId(Long studentId);
     Student findByUserUsername(User u);
